Fix findCabalFileDir: dir should be prepended to all filenames
In an unlikely event of several Cabal file in the same folder
Stan currently chooses to return the first of them, but inadverently
forgets to prepend a directory name.

Merging (x : _xs) case with [cabalFile] has a nice side effect
of making the search a bit faster: previous version of the code
would necessarily scan the entire folder looking for the second
Cabal file (because if it is to be found, outcomes will be different),
but now we shall stop after finding the first one.

Related to https://github.com/haskell/haskell-language-server/issues/4515