Bump `stack.yaml` to GHC 9.10.2/lts-24.0
Also bumps old `stack.yaml` to lts-23.27 (the last for GHC 9.8.4) and preserves it, for people using Stack with GHC 9.8.4 and Stack in CI.

Also adds Stack project-level configuration file for GHC 9.6.7.

Also revisits the Stack project-level configuration files for GHC 9.4.8, 9.6.3 to 9.6.6.

Also bumps the Stack-related part of the CI.

Also bumps the non-Stack related part of the CI.